requirements.txt
cryptography==41.0.3
argon2-cffi==21.3.0
pytest==7.4.0
